+++ src/main/data/rasputin.txt
@@ -1,4 +1,4 @@
-D|1|return book|2024-08-29 1800
+D|0|return book|2024-08-29 1800
 E|1|project meeting|2024-08-28 1400|2024-08-28 1500
 T|1|read book
 D|0|update calendar|2024-08-30
