+++ src/main/java/Command/Command.java
@@ -1,3 +1,7 @@
+package Command;
+
+import Task.RasputinException;
+
 public abstract class Command {
 
     public abstract void execute() throws RasputinException;
