+++ src/main/java/Command/UnmarkCommand.java
@@ -1,9 +1,16 @@
-public class DeleteCommand extends Command {
+package Command;
+
+import Task.TaskList;
+import Task.InvalidTaskException;
+
+import Ui.Ui;
+
+public class UnmarkCommand extends Command {
 
     private TaskList tasks;
     private int index;
 
-    public DeleteCommand(TaskList tasks, int index) {
+    public UnmarkCommand(TaskList tasks, int index) {
         this.tasks = tasks;
         this.index = index;
     }
@@ -11,8 +18,8 @@ public class DeleteCommand extends Command {
     @Override
     public void execute() {
         try {
-            Ui.printText("Done, removed that task for you.\n" + tasks.get(index).toString());
-            tasks.remove(index);
+            tasks.get(index).markAsNotDone();
+            Ui.printText("Task has been unmarked.\n" + tasks.get(index).toString());
         } catch (IndexOutOfBoundsException e) {
             throw new InvalidTaskException("ERROR! Task not found.");
         }
