+++ src/main/java/Task/InvalidTaskException.java
@@ -1,3 +1,5 @@
+package Task;
+
 public class InvalidTaskException extends RasputinException {
 
     public InvalidTaskException(String message) {
