+++ src/main/java/Task/RasputinException.java
@@ -1,3 +1,5 @@
+package Task;
+
 public class RasputinException extends RuntimeException {
     public RasputinException(String message) {
         super(message);
