+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
-Task file not found, creating task file.
+Task.Task file not found, creating task file.
 Unable to create task file.
-Task file not found.
+Task.Task file not found.
 ____________________________________
 Hello, I'm Rasputin!
 What can I do for you?
